HistoryHash merged into SpreadsheetHistoryHash
- Also fixes several warnings at runtime due to invalid Prop definitions for history props.
